refactor: feedListSkeleton  리팩토링
